requirements.txt
numpy == 1.19.3
pandas == 1.1.3
pyyaml
munch
binance-futures-connector
pytest
requests
flask ~= 2.2.2
flask-cors ~= 3.0.10
flask-sqlalchemy ~= 3.0.2
SQLAlchemy ~= 1.4.46
TA-Lib
